Move SystemTestDrivers to base package
