+++ java/system-test/system-test-dsl/src/main/java/sleeper/systemtest/dsl/SystemTestDrivers.java
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package sleeper.systemtest.dsl.util;
+package sleeper.systemtest.dsl;
 
-import sleeper.systemtest.dsl.SystemTestContext;
 import sleeper.systemtest.dsl.compaction.CompactionDriver;
 import sleeper.systemtest.dsl.ingest.DirectBulkImportDriver;
 import sleeper.systemtest.dsl.ingest.DirectIngestDriver;
@@ -41,6 +40,8 @@ import sleeper.systemtest.dsl.reporting.PartitionReportDriver;
 import sleeper.systemtest.dsl.sourcedata.DataGenerationTasksDriver;
 import sleeper.systemtest.dsl.sourcedata.GeneratedIngestSourceFilesDriver;
 import sleeper.systemtest.dsl.sourcedata.IngestSourceFilesDriver;
+import sleeper.systemtest.dsl.util.PurgeQueueDriver;
+import sleeper.systemtest.dsl.util.WaitForJobs;
 
 public interface SystemTestDrivers {
