+++ java/system-test/system-test-dsl/src/test/java/sleeper/systemtest/dsl/extension/AfterTestReports.java
@@ -17,8 +17,8 @@
 package sleeper.systemtest.dsl.extension;
 
 import sleeper.systemtest.dsl.SystemTestContext;
+import sleeper.systemtest.dsl.SystemTestDrivers;
 import sleeper.systemtest.dsl.reporting.SystemTestReports;
-import sleeper.systemtest.dsl.util.SystemTestDrivers;
 
 public class AfterTestReports extends AfterTestReportsBase<SystemTestReports.SystemTestBuilder> {
